Reformated parameter and variable files per Pavel's suggestions. Wrote new tests.
