+++ policyengine_us/parameters/gov/states/ny/nyserda/drive_clean/amount.yaml
@@ -1,4 +1,4 @@
-description: The Drive Clean program offers a different rebate based on the all-electric range of the vehicle and the date of its purchase.
+description: New York provides the following Drive Clean point-of-sale rebate amount, based on the all-electric range of the vehicle.
 brackets:
   # 0 - 19 miles 
   - threshold:
