--- policyengine_us/parameters/gov/states/ny/nyserda/drive_clean/drive_clean_high_msrp.yaml
@@ -1,12 +1,0 @@
-description: New York minimizes the Drive Clean point-of-sale rebate for high-price vehicles.
-metadata:
-  unit: miles
-  period: year
-  label: High Manufacturer's Suggested Price
-  reference:
-    - title: NYSERDA Drive Clean Implementation Manual
-      href: https://www.nyserda.ny.gov/All-Programs/Drive-Clean-Rebate-For-Electric-Cars-Program/Dealers/Dealer-Resources
-
-values:
-  2017-03-21: 60_000
-  2021-06-30: 42_000
