+++ policyengine_us/parameters/gov/states/ny/nyserda/drive_clean/flat_rebate/amount.yaml
@@ -1,8 +1,7 @@
-description: New York minimizes the Drive Clean point-of-sale rebate for high-price vehicles.
-metadata:
+description: New York provides the following flat Drive Clean point-of-sale rebate vehicles with a base MSRP over a certain threshold.
   unit: currency-usd
   period: year
-  label: High Manufacturer's Suggested Price
+  label: New York Drive Clean point-of-sale flat rebate MSRP amount
   reference:
     - title: NYSERDA Drive Clean Implementation Manual
       href: https://www.nyserda.ny.gov/All-Programs/Drive-Clean-Rebate-For-Electric-Cars-Program/Dealers/Dealer-Resources
