+++ policyengine_us/parameters/gov/states/ny/nyserda/drive_clean/flat_rebate/msrp_threshold.yaml
@@ -1,0 +1,12 @@
+description: New York provides a flat Drive Clean point-of-sale rebate for vehicles with a base MSRP over this amount.
+metadata:
+  unit: miles
+  period: year
+  label: New York Drive Clean point-of-sale flat rebate MSRP threshold
+  reference:
+    - title: NYSERDA Drive Clean Implementation Manual
+      href: "https://www.nyserda.ny.gov/-/media/Project/Nyserda/Files/Programs/Drive-Clean-NY/implementation-manual.pdf" #page=8
+
+values:
+  2017-03-21: 60_000
+  2021-06-30: 42_000
